Improve Queries and datatable

+++ db/migrations/20240519230248_create_tournament_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE tournaments (
     id BIGSERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    entry_fee INT NOT NULL,
+    entry_fee BIGINT NOT NULL,
     start_time TIMESTAMP NOT NULL
 );
